remove_confirm_modal is migrated
